backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.34.0
httpx==0.27.0
python-dotenv==1.0.0
mlflow==2.15.0
aerospike==15.1.0
pandas==2.2.2
pydantic==2.8.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
overmind-sdk
synix @ git+https://github.com/marklubin/synix.git
